requirements.txt
Flask==3.0.3
pandas==1.5.3
pandasai==2.0.37
python-dotenv==1.0.1
redis==5.0.4
